moved others suggested by Keith
