+++ eidoscommon/src/main/scala/org/clulab/wm/eidoscommon/utils/Counter.scala
@@ -1,4 +1,4 @@
-package org.clulab.wm.eidos.utils
+package org.clulab.wm.eidoscommon.utils
 
 class Counter(protected var value: Int, step: Int) {
