+++ eidoscommon/src/main/scala/org/clulab/wm/eidoscommon/utils/EidosException.scala
@@ -1,4 +1,4 @@
-package org.clulab.wm.eidos.utils
+package org.clulab.wm.eidoscommon.utils
 
 class EidosException(message: String, cause: Throwable) extends RuntimeException(message, cause) {
