+++ eidoscommon/src/main/scala/org/clulab/wm/eidoscommon/utils/First.scala
@@ -1,4 +1,4 @@
-package org.clulab.wm.eidos.utils
+package org.clulab.wm.eidoscommon.utils
 
 class First {
   protected var first = true
