+++ src/main/scala/org/clulab/wm/eidos/utils/meta/PlainText.scala
@@ -4,10 +4,10 @@ import java.text.SimpleDateFormat
 import java.util.Calendar
 import java.util.TimeZone
 
-import org.clulab.wm.eidos.utils.EidosException
 import org.clulab.timenorm.scate.SimpleInterval
 import org.clulab.wm.eidos.context.DCT
 import org.clulab.wm.eidos.document.Metadata
+import org.clulab.wm.eidoscommon.utils.EidosException
 
 class PlainText(text: String,
   titleOpt: Option[String] = None,
