+++ src/test/scala/org/clulab/wm/eidos/utils/TestQuicklyEqualable.scala
@@ -2,6 +2,7 @@
 package org.clulab.wm.eidos.utils
 
 import org.clulab.wm.eidos.test.TestUtils._
+import org.clulab.wm.eidoscommon.utils.QuicklyEqualable
 
 class TestQuicklyEqualable extends Test {
